メモ機能実装

+++ app/controllers/chats_controller.rb
@@ -2,6 +2,11 @@ class ChatsController < ApplicationController
   before_action :set_room
 
   def index
+    if @room.memos.find_by(user_id: current_user.id).present?
+      @memo = @room.memos.find_by(user_id: current_user.id)
+    else
+      @memo = Memo.new
+    end
     @question_sheet = QuestionSheet.new
     5.times { @question_sheet.options.build }
     @archive = Archive.new
@@ -18,21 +23,25 @@ class ChatsController < ApplicationController
       redirect_to room_chats_path(@room)
       # render json: { chat: @chat }
     else
-      redirect_to room_chats_path(@room)
+      render :index
     end
   end
 
   def destroy
     chat = Chat.find(params[:id])
-    chat.destroy
-    redirect_to room_chats_path(@room)
+    if chat
+      chat.destroy
+      redirect_to room_chats_path(@room), notice: 'Chat was successfully destroyed.'
+    else
+      redirect_to room_chats_path(@room), alert: 'Chat not found.'
+    end
   end
 
   private
 
   def set_room
     @room = Room.find(params[:room_id])
-  end
+   end
 
   def chat_params
     params.require(:chat).permit(:message, :image).merge(user_id: current_user.id)
